feat: new miniature view
